Refactored Tag Logic
Much more robust implementation where all tag values are stored and sent as opposed to relying on individual checkbox event handlers
