--- src/models/TagCategory.ts
@@ -1,5 +1,0 @@
-export default interface TagCategory {
-    tag: string;
-    label: string;
-    checked?: boolean;
-}
