+++ src/models/Tags.ts
@@ -1,0 +1,9 @@
+export interface TagCategory {
+    tag: string;
+    label: string;
+    checked?: boolean;
+}
+
+export interface TagsObject {
+    [key:string]: string|boolean|number;
+}
